# docs/swagger.yaml
basePath: /v1
definitions:
  httpx.DataJsonResponse:
    properties:
      data: {}
    type: object
  httpx.ErrorJsonResponse:
    properties:
      error:
        type: string
    type: object
  httpx.MessageJsonResponse:
    properties:
      message:
        type: string
    type: object
...
  schemas.QuestionResponse:
    properties:
      id:
        type: integer
      max_text:
        type: string
      min_text:
        type: string
      number:
        type: integer
      options:
        items:
          type: string
        type: array
      question:
        type: string
      type:
        type: string
    type: object
  schemas.QuestionnareSettingAlternative:
    properties:
      alternative:
        type: string
      entrepreneurship_opportunity:
        type: integer
      id:
        type: integer
      salary:
        type: integer
      total_open_jobs:
        type: integer
    type: object
  schemas.QuestionnareSettingRequest:
    properties:
      alternative_id:
        type: integer
      entrepreneurship_opportunity:
        type: integer
      salary:
        type: integer
      total_open_jobs:
        type: integer
    type: object
  schemas.QuestionnareStatusResponse:
    properties:
      status:
        type: string
    type: object
...
  schemas.SubmitAnswerRequest:
    properties:
      answer:
        type: string
      id:
        type: integer
      number:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a Double Track REST API
  title: Double Track API
  version: "1.0"
paths:
...
  /questionnare/answers:
    delete:
      consumes:
      - application/json
      description: Delete student questionnare answer
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpx.MessageJsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Delete student questionnare answer
      tags:
      - Questionnare
      - Student
    post:
      consumes:
      - application/json
      description: Submit questionnare answers
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Submit answer request body
        in: body
        name: body
        required: true
        schema:
          items:
            $ref: '#/definitions/schemas.SubmitAnswerRequest'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpx.MessageJsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Submit answer
      tags:
      - Questionnare
      - Student
  /questionnare/questions:
    get:
      consumes:
      - application/json
      description: Get all available questions
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.DataJsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schemas.QuestionResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Get Questions
      tags:
      - Questionnare
      - Student
  /questionnare/settings:
    get:
      consumes:
      - application/json
      description: Get questionnare settings
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.DataJsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schemas.QuestionnareSettingAlternative'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Get questionnare settings
      tags:
      - Questionnare
      - Admin
    post:
      consumes:
      - application/json
      description: Add a questionnare setting
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Add questionnare setting request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schemas.QuestionnareSettingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpx.MessageJsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Add questionnare setting
      tags:
      - Questionnare
      - Admin
  /questionnare/settings/incomplete:
    get:
      consumes:
      - application/json
      description: Get incomplete questionnare settings
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.DataJsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schemas.QuestionnareSettingAlternative'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Get incomplete questionnare settings
      tags:
      - Questionnare
      - Admin
  /questionnare/status:
    get:
      consumes:
      - application/json
      description: Get questionnare status
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.DataJsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/schemas.QuestionnareStatusResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Get questionnare status
      tags:
      - Questionnare
      - Student
...
swagger: "2.0"